+++ apps/web-antd/src/plugins/detective/utils/echarts.ts
@@ -1,0 +1,27 @@
+/**
+ * ECharts 按需导入模块
+ * 仅导入项目中使用的 bar / pie / line 图表及必要组件
+ */
+import { BarChart, LineChart, PieChart } from 'echarts/charts';
+import {
+  GridComponent,
+  LegendComponent,
+  TitleComponent,
+  TooltipComponent,
+} from 'echarts/components';
+import * as echarts from 'echarts/core';
+import { CanvasRenderer } from 'echarts/renderers';
+
+echarts.use([
+  BarChart,
+  LineChart,
+  PieChart,
+  GridComponent,
+  TooltipComponent,
+  LegendComponent,
+  TitleComponent,
+  CanvasRenderer,
+]);
+
+export default echarts;
+export type { EChartsOption } from 'echarts';